Version 1 Example Level Prototype
Version 1.0:
Some elements (such as food 'pick-ups', UI text elements, etc.) may be missing from this current version.
Some elements such as the background, grid, UI texts, etc. aren't finalised and will require feedback and/or alteration.
A couple elements such as the buttons on the top right, the 'turn counter' in the bottom right might be finalised but again, requires confirmation from the team as a whole.